refactor: rename useShareState to useState
