+++ packages/entry/src/index.js
@@ -15,6 +15,6 @@ export { useCompile } from './templateHash'
 export { defineEntry, callEntry, beforeCallEntry, afterCallEntry, mergeRegistry, getMergeRegistry } from './entryHash'
 export { getLayoutComponent } from './layoutHash'
 export { default as useMessage } from './useMessage'
-export { useShareState, useStore } from './useShareState'
+export { useState, useStore } from './useState'
 export { getConfigurator, addConfigurator } from './configurators'
 export * from './hooks'
